🐛 fix small bugs on number formatting and contracts

+++ app/Models/InvoiceItem.php
@@ -37,10 +37,10 @@ class InvoiceItem extends Model implements Auditable
     {
         return [
             'quantity' => 'integer',
-            'price' => 'float',
-            'discount' => 'float',
-            'tax' => 'float',
-            'total' => 'float',
+            'price' => 'double',
+            'discount' => 'double',
+            'tax' => 'double',
+            'total' => 'double',
         ];
     }
